+++ contrib/electrum-plugin/bwt.py
@@ -138,13 +138,15 @@ class BwtPlugin(BasePlugin):
 
     @hook
     def load_wallet(self, wallet, main_window):
-        if wallet.get_master_public_keys():
+        if not wallet.get_master_public_keys():
+            _logger.warning('skipping unsupported wallet type %s' % wallet.wallet_type)
+        elif not wallet.is_watching_only():
+            _logger.warning('skipping hot wallet')
+        else:
             num_wallets = len(self.wallets)
             self.wallets |= {wallet}
             if len(self.wallets) != num_wallets:
                 self.start()
-        else:
-            _logger.warning('%s wallets are unsupported, skipping' % wallet.wallet_type)
 
     @hook
     def close_wallet(self, wallet):